refactored_pnl
