+++ ta4j-core/src/main/java/org/ta4j/core/reports/PerformanceReportGenerator.java
@@ -26,10 +26,10 @@ package org.ta4j.core.reports;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.ProfitLossCriterion;
-import org.ta4j.core.analysis.criteria.ProfitLossPercentageCriterion;
-import org.ta4j.core.analysis.criteria.TotalLossCriterion;
-import org.ta4j.core.analysis.criteria.TotalProfitCriterion;
+import org.ta4j.core.analysis.criteria.pnl.NetLossCriterion;
+import org.ta4j.core.analysis.criteria.pnl.NetProfitCriterion;
+import org.ta4j.core.analysis.criteria.pnl.ProfitLossCriterion;
+import org.ta4j.core.analysis.criteria.pnl.ProfitLossPercentageCriterion;
 import org.ta4j.core.num.Num;
 
 /**
@@ -42,10 +42,10 @@ public class PerformanceReportGenerator implements ReportGenerator<PerformanceRe
 
     @Override
     public PerformanceReport generate(Strategy strategy, TradingRecord tradingRecord, BarSeries series) {
-        final Num totalProfitLoss = new ProfitLossCriterion().calculate(series, tradingRecord);
-        final Num totalProfitLossPercentage = new ProfitLossPercentageCriterion().calculate(series, tradingRecord);
-        final Num totalProfit = new TotalProfitCriterion().calculate(series, tradingRecord);
-        final Num totalLoss = new TotalLossCriterion().calculate(series, tradingRecord);
-        return new PerformanceReport(totalProfitLoss, totalProfitLossPercentage, totalProfit, totalLoss);
+        final Num pnl = new ProfitLossCriterion().calculate(series, tradingRecord);
+        final Num pnlPercentage = new ProfitLossPercentageCriterion().calculate(series, tradingRecord);
+        final Num netProfit = new NetProfitCriterion().calculate(series, tradingRecord);
+        final Num netLoss = new NetLossCriterion().calculate(series, tradingRecord);
+        return new PerformanceReport(pnl, pnlPercentage, netProfit, netLoss);
     }
 }
