+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/GrossProfitCriterionTest.java
@@ -21,17 +21,18 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.analysis.criteria;
+package org.ta4j.core.analysis.criteria.pnl;
 
 import java.util.function.Function;
 
 import org.junit.Test;
+import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
 import org.ta4j.core.num.Num;
 
-public class TotalProfitCriterionTest extends AbstractCriterionTest {
+public class GrossProfitCriterionTest extends AbstractCriterionTest {
 
-    public TotalProfitCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new TotalProfitCriterion(), numFunction);
+    public GrossProfitCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new GrossProfitCriterion(), numFunction);
     }
 
     @Test
