+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/NetLossCriterionTest.java
@@ -21,7 +21,7 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.analysis.criteria;
+package org.ta4j.core.analysis.criteria.pnl;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -34,13 +34,14 @@ import org.ta4j.core.AnalysisCriterion;
 import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
+import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
-public class TotalLossCriterionTest extends AbstractCriterionTest {
+public class NetLossCriterionTest extends AbstractCriterionTest {
 
-    public TotalLossCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new TotalLossCriterion(), numFunction);
+    public NetLossCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new NetLossCriterion(), numFunction);
     }
 
     @Test
