+++ ta4j-examples/src/main/java/ta4jexamples/num/CompareNumTypes.java
@@ -33,7 +33,7 @@ import org.ta4j.core.BaseStrategy;
 import org.ta4j.core.Rule;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.TotalReturnCriterion;
+import org.ta4j.core.analysis.criteria.pnl.GrossReturnCriterion;
 import org.ta4j.core.indicators.EMAIndicator;
 import org.ta4j.core.indicators.MACDIndicator;
 import org.ta4j.core.indicators.RSIIndicator;
@@ -91,7 +91,7 @@ public class CompareNumTypes {
         long start = System.currentTimeMillis();
         BarSeriesManager manager = new BarSeriesManager(series);
         TradingRecord record1 = manager.run(strategy1);
-        TotalReturnCriterion totalReturn1 = new TotalReturnCriterion();
+        GrossReturnCriterion totalReturn1 = new GrossReturnCriterion();
         Num returnResult1 = totalReturn1.calculate(series, record1);
         long end = System.currentTimeMillis();
