+++ ta4j-examples/src/main/java/ta4jexamples/strategies/MovingMomentumStrategy.java
@@ -29,7 +29,7 @@ import org.ta4j.core.BaseStrategy;
 import org.ta4j.core.Rule;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.TotalReturnCriterion;
+import org.ta4j.core.analysis.criteria.pnl.GrossReturnCriterion;
 import org.ta4j.core.indicators.EMAIndicator;
 import org.ta4j.core.indicators.MACDIndicator;
 import org.ta4j.core.indicators.StochasticOscillatorKIndicator;
@@ -101,6 +101,6 @@ public class MovingMomentumStrategy {
 
         // Analysis
         System.out.println(
-                "Total profit for the strategy: " + new TotalReturnCriterion().calculate(series, tradingRecord));
+                "Total profit for the strategy: " + new GrossReturnCriterion().calculate(series, tradingRecord));
     }
 }
