+++ ta4j-examples/src/main/java/ta4jexamples/strategies/RSI2Strategy.java
@@ -29,7 +29,7 @@ import org.ta4j.core.BaseStrategy;
 import org.ta4j.core.Rule;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.TotalReturnCriterion;
+import org.ta4j.core.analysis.criteria.pnl.GrossReturnCriterion;
 import org.ta4j.core.indicators.RSIIndicator;
 import org.ta4j.core.indicators.SMAIndicator;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
@@ -98,7 +98,7 @@ public class RSI2Strategy {
 
         // Analysis
         System.out.println(
-                "Total return for the strategy: " + new TotalReturnCriterion().calculate(series, tradingRecord));
+                "Total return for the strategy: " + new GrossReturnCriterion().calculate(series, tradingRecord));
     }
 
 }
